Add files via upload
The purpose of this project is to create an enrollment procedure that checks for multiple variables before determining whether or not the student is qualified to enroll into a class. It also checks whether or not a class is full nor not and if so, enrolls the applying student into a WaitList table.